[REFACTOR] Moved files, removed a lot of useless types (#33)
* fix: Moved `wezterm.` submodules out of their directories.


* refactor!: Dropped some aliases that were useless.


* fix(config): Moved `types/config/init.lua` to `types/config.lua`.


* chore: Fixed and updated CODEOWNERS lines.


---------

+++ lua/types/objects/color.lua
@@ -76,7 +76,7 @@
 ---@field square fun(self: Color): a: Color, b: Color, c: Color
 -- Returns a tuple of the internal `SRGBA` colors expressed
 -- as unsigned 8-bit integers in the range `0-255`
----@field srgb_u8 fun(self: Color): r: u8, g: u8, b: u8, alpha: u8
+---@field srgb_u8 fun(self: Color): r: integer, g: integer, b: integer, alpha: integer
 -- Returns the other two colors that form a triad.
 -- The other colors are at +/- 120 degrees in the `HSL` color wheel
 ---@field triad fun(self: Color): a: Color, b: Color
